requirements.txt
omegaconf==2.3.0
numpy==1.25.1
pandas==2.0.3
pydantic==1.10.11
torch==2.0.1
timm==0.9.2
torchmetrics==1.0.1
scikit-learn==1.3.0
fastapi==0.103.1
uvicorn==0.23.2
Pillow==10.0.0
torchvision==0.15.2